backend/requirements.txt
supabase
flask
openai
flasgger
python-dotenv
numpy